requirements.txt
akshare
yfinance
pandas
numpy
requests
streamlit
openai
pyyaml
schedule
aiohttp
pandas-ta
plotly
croniter
colorama
python-dateutil